perf: collect changed DOM elements efficiently
Previously, the whole DOM was serialized and then just filtered by
previously sent elements. Now the virtual DOM marks elements and
subtrees as changed, which allows to efficiently iterate only over the
elements that changed since the last update.

The virtual DOM itself was already updated incrementally, just the
synchronization to the client was a temporary inefficient
implementation.