fix(react): generate java lambdas instead of anonymous classes
MPS can only generate closures to lambdas if some conditions are meet.
Especially duplicate variable names are not allowed.
Generating hundreds of anonymous classes results in a poor performance of the compiler.
Lambdas are now used directly in the generator instead of closures to prevent a regression.
The compiler will then complain if one of the conditions is not meet.